// chat-ui/redux/action-types.ts
export enum ReduxActionTypes {
  SET_LOGGED_IN = "SET_LOGGED_IN",
  SET_USER_INFO = "SET_USER_INFO",
  SET_SOCKET_CONNECTED = "SET_SOCKET_CONNECTED",
  SET_CHATS = "SET_CHATS",
  UPDATE_CHATS = "UPDATE_CHATS",
  UPDATE_A_CHAT = "UPDATE_A_CHAT",
  SET_NEW_MESSAGE = "SET_NEW_MESSAGE",
  SEND_NEW_MESSAGE = "SEND_NEW_MESSAGE",
  SET_NEW_ACTIVE_CHAT = "SET_NEW_ACTIVE_CHAT",
  CLEAR_UNREAD_COUNT_CHAT = "CLEAR_UNREAD_COUNT_CHAT",
  LOG_OUT = "LOG_OUT",
}